requirements.txt
discord
dotenv
google.generativeai
mysql-connector-python
firebase_admin
yt_dlp
